requirements.txt
dynaconf==3.2.12
beautifulsoup4==4.14.3
requests==2.32.5
google-genai==1.63.0
